fix: scope no-unsupported-api exemption to only the 1.13.0+ code
Disabling obsidianmd/no-unsupported-api for all of settings.ts meant
any future edit anywhere else in that file (display(), the settings
model, etc.) could accidentally use an API requiring a newer
minAppVersion than 1.8.7 without the linter catching it -- inline
eslint-disable comments aren't an option here, since this repo's own
eslint-comments/no-restricted-disable rule explicitly blocks
disabling any obsidianmd/* rule that way.

Moved the four Obsidian-1.13.0+-only methods (getSettingDefinitions,
buildWorkspacePage, buildModePage, getControlValue, setControlValue,
plus the parseOverrideKey helper) into a new settingsDeclarative.ts,
called from thin delegator methods on WorkspacesPlusSettingsTab.
Replaced the two `super.getControlValue`/`super.setControlValue`
calls (which can only be written inside the subclass's own class
body, so couldn't be moved by themselves) with
`PluginSettingTab.prototype.getControlValue.call(tab, key)` and the
setControlValue equivalent -- functionally identical to `super.X()`
here, but expressible from a plain function in the new file instead
of requiring the `super` keyword.

The eslint.config.mjs override now targets only settingsDeclarative.ts.
settings.ts no longer references any 1.13.0+-only API at all, so it
needs no exemption -- the check is fully active for display() and
the rest of the file again.

TOGGLE_TEXT and getChildIds are now exported from settings.ts for
settingsDeclarative.ts to import; this does introduce a circular
module reference between the two files (Rollup warns but builds
fine), safe here since every cross-reference happens inside function
bodies rather than at module-evaluation time.
